Fix broken documentation link in README
This commit updates the link to the commands.md file in the README to
reflect the new directory structure. The documentation was moved from
the 'doc' directory to 'docs' as part of our improved documentation
organization, but the README reference wasn't updated accordingly.

This simple fix ensures users can easily find the command reference
documentation when navigating from the README.